Use normal UART on H3 devices
Currently ar100-info is using the port F UART (accessible via
the SD breakout board on the SD card slot). But it is more
convenient to use just a normal UART on popular Allwinner H3
development boards, such as Orange Pi PC.

This patch is borrowing some constant definitions from U-Boot,
which are needed to initialize the UART hardware.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>