Move the stack up and get 32K for code instead of only 16K
The commit e5096906b42048a7f32259403f4fed3cfde14b2b moved the
stack down by 32K with the intention to support Allwinner H3,
which only has 32K of SRAM instead of 64K in Allwinner A31.

But because of the large reset vectors padding area in the
beginning of the address space, the address/size arithmetics
is just a little bit more tricky and in fact ar100-info
ended up with the usable space reduced to only 16K (down
from 48K). Oops.

Now move the stack pointer to the real end of SRAM and get
full 32K of usable space on Allwinner H3.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>