Add simple benchmark for SRAM A1 and SRAM A2
The OpenRISC core has access to at last two SRAM areas: A1 and A2.
SRAM A1 is primarily used by the ARM cores and SRAM A2 is closely
integrated with the OpenRISC core (see the "System Bus Tree"
diagram in the Allwinner H3 datasheet). As a result, SRAM A2 is
very fast but SRAM A1 is relatively slow.

This patch adds a simple benchmark for fetching instructions
and reading data from SRAM with and without I-cache. Up to date
results should be available in the linux-sunxi wiki:

    https://linux-sunxi.org/AR100

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>