Do not initialize DRAM on Allwinner A31 if it is already clocked
The current DRAM init code from ar100-info does not work on
my Allwinner A31s tablet. One of the possible problems is the
number of channels (two in A31 vs. one in A31s) but quickly
changing the number of channels to 1 still does not help and
the DRAM initialization still gets stuck.

So treat A31/A31s in the same way as H3 and skip the DRAM
initialization if the DRAM is already clocked (the DRAM can
be already initialized by the U-Boot SPL).

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>